Propose special ReceiptCourier interface
This PR proposes using a separate interface, `ReceiptCourier` for
defining couriers that are capable of giving a receipt for a specific
email. This allows the simple Courier interface to still work for local
delivery mechanisms but still supports getting the receipt ID for any
given email on SaaS delivery mechanisms.