Move to a logging-based courier for local testing (#35)
It was presented that the MailCourier could introduce injection
attacks into a developer's system based on the current implementation.
Because the courier was built as a way to easily run tests locally,
rather than as a production-ready system, it is being replaced with a
courier that uses a LoggerInterface to document the email instead. This
gives developers a solution to put into their local environments for
testing and debugging, reducing the possibility of it being misused in a
production system.

I would like to revist the MailCourier at a later time, possibly as a
third-party adapter instead of built into the core package itself.